Move filterAndSortPeriodicalEditions out of find on shelf dropdowns
This functionality will need to be reused outside of the component on
the material page for periodical dropdowns. It also cleans up the
"find on shelf periodical dropdowns" component and makes it easier to
read.
